Enable CLANG_WARN_IMPLICIT_SIGN_CONVERSION

+++ Source/OCMockito/Invocation/NSInvocation+OCMockito.m
@@ -21,7 +21,7 @@
     for (NSUInteger idx = 2; idx < numberOfArguments; ++idx) // Indices 0 and 1 are self and _cmd
     {
         const char *argType = [signature getArgumentTypeAtIndex:idx];
-        id arg = [MKTArgumentGetterChain() retrieveArgumentAtIndex:idx ofType:argType onInvocation:self];
+        id arg = [MKTArgumentGetterChain() retrieveArgumentAtIndex:(NSInteger)idx ofType:argType onInvocation:self];
         if (arg)
             [arguments addObject:arg];
         else
